Move sunburst up and remove dummy data
