--- src/components/Sunburst/SunburstChart/data.json
@@ -1,116 +1,0 @@
-{
-  "name": "Partners",
-  "children": [
-    {
-      "name": "Vlaio",
-      "children": [
-        {
-          "name": "NSZ",
-          "children": [
-            {
-              "name": "UGent",
-              "size": 41
-            },
-            {
-              "name": "KULeuven",
-              "size": 4
-            }
-          ]
-        },
-        {
-          "name": "Voka",
-          "children": [
-            {
-              "name": "Vlaio",
-              "size": 200
-            },
-            {
-              "name": "KULeuven",
-              "size": 12
-            }
-          ]
-        },
-        {
-          "name": "UGent",
-          "size": 4
-        },
-        {
-          "name": "Unizo",
-          "size": 50
-        }
-      ]
-    },
-    {
-      "name": "KULeuven",
-      "children": [
-        {
-          "name": "UGent",
-          "size": 120
-        },
-        {
-          "name": "Unizo",
-          "size": 50
-        },
-        {
-          "name": "NSZ",
-          "children": [
-            {
-              "name": "Voka",
-              "children": [
-                {
-                  "name": "Vlaio",
-                  "children": [
-                    {
-                      "name": "NSZ",
-                      "size": 15
-                    }
-                  ]
-                },
-                {
-                  "name": "KULeuven",
-                  "children": [
-                    {
-                      "name": "UGent",
-                      "size": 20
-                    }
-                  ]
-                }
-              ]
-            },
-            {
-              "name": "KULeuven",
-              "size": 12
-            }
-          ]
-        },
-        {
-          "name": "Voka",
-          "children": [
-            {
-              "name": "NSZ",
-              "size": 150
-            }
-          ]
-        },
-        {
-          "name": "Vlaio",
-          "children": [
-            {
-              "name": "NSZ",
-              "size": 65
-            }
-          ]
-        }
-      ]
-    },
-    {
-      "name": "UGent",
-      "children": [
-        {
-          "name": "NSZ",
-          "size": 80
-        }
-      ]
-    }
-  ]
-}
